feat: add MATH benchmark
